moved bench_* files into subdir; minor logging clean up
