+++ benchmark/benchmark/bench_kmeans.py
